pyproject.toml
[project]
name = "praicing"
version = "0.0.0"
description = "Estimate the costs of running benchmarks across different model providers."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "João Palmeiro", email = "joaopalmeiro@proton.me" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = ["pillow>=11.0.0"]

[project.urls]
Issues = "https://github.com/joaopalmeiro/praicing/issues"
Source = "https://github.com/joaopalmeiro/praicing"

[dependency-groups]
dev = ["mypy[faster-cache]==1.16.1", "ruff==0.11.2"]

[build-system]
requires = ["hatchling==1.27.0"]
build-backend = "hatchling.build"
